Disconnect Ch376 interrupt
Do not connect Ch376 interrupt line to the CPU INT pin; use Ch376
GET_STATUS command to read the interrupt status in UART mode.
If necessary, a separate register would be better to read this
interrupt.

Keep the three oscillators always running in the RNG register chip. If
they are turned on only when a value is requested, the results are not
very random.